Removed redundant and confusing sentence
Remove a sentence that is not only redundant but possibly confusing
since it can be read in a way that contradicts the previous paragraph.
Specifically, the previous paragraph says that root must be sda2
whereas this sentence suggests that the partitioning layout is
completely arbitrary.

Thanks to Alexander Zenger for the feedback.