Document u-boot upgrade process on OpenRD
Debian provides mainline u-boot images for OpenRD now, so describe
how the upgrade process to give users the option to use a modern
u-boot.

Note that we use 0x60000 as the maximum u-boot filesize.  We cannot
use ${filesize} because the original Marvell u-boot doesn't handle
that properly.

Thanks to Rick Thomas for testing u-boot on OpenRD and to Albert
ARIBAUD for maintaining it upstream!