Remove the use of <a id>
As I was discussing a problem I saw in the way <h> elements were
formatted (headers with an `<a id>` were blue with an underline
when you hover over it whereas othe headers were black), Chris
Warrick from the Nikola project pointed out that my use of
`<h2><a id = "...">...</a></h2>` is "something one would do in
the 90s".  In fact, `<h2 id="...>` solves the problem I was seeing.